Added files to appropriate folders
