+++ HashMap/hashmap.java
