+++ HashMap/remove_adjacent_duplicates.java
