+++ Queue/queue_byLinkedlist.java
@@ -1,36 +1,57 @@
-package Queue;
-
-public class queue_byLinkedlist {
-    public static void main(String[] args) {
-
-    }
-    public class Node{
+public class Queue_byLinkedList {
+    static class Node{
         int data;
         Node next;
-        public Node(int data){
+        Node(int data){
             this.data = data;
+            next = null;
         }
     }
-    public class Queue{
-        Node front,rear;
-        void enqueue(int data){
+    static class Queue{
+        static Node head = null;
+        static Node tail = null;
+
+        public static boolean isEmpty(){
+            return head == null && tail == null;
+        }
+         public static void add(int data){
             Node newNode = new Node(data);
-        if (front == null ){
-            front=rear=newNode;
-            return;
-        }
-        rear.next = newNode;
-        rear = newNode;
-        }
-
-        int dequeue() throws Exception {
-            if (front == null){
-                throw new Exception();
+            if(tail == null) {
+                tail=head=newNode;
+                return;
             }
-            int result = front.data;
-            front = front.next;
-            return result;
+            tail.next = newNode;
+            tail = newNode; // assigning newNode as value to tail
+         }
+         public static int remove(){
+            if (isEmpty()){
+                System.out.println("empty queue");
+                return -1;
+            }
+            int front = head.data;
+            if(head == tail) tail = null;
+
+            head = head.next;
+            return front;
+         }
+         public static int peek(){
+            if (isEmpty()){
+                System.out.println("empty queue");
+                return -1;
+            }
+            return head.data;
+         }
+    }
+    public static void main(String[] args) {
+        Queue q = new Queue();
+        q.add(0);
+        q.add(1);
+        q.add(2);
+        q.add(3);
+        //q.remove();
+        while(!q.isEmpty()){
+            System.out.println(q.peek());
+            q.remove();
         }
     }
-
 }
