+++ README.md
@@ -1,13 +1,16 @@
 # Data Structures and Algorithms in Java
+
 This is a collection of my implementations of various Data Structures and Algorithms in Java. These implementations are intended to help me learn and practice the concepts of Data Structures and Algorithms.
 
 ## Table of Contents
+
 - Data Structures
 - Algorithms
 - Usage
 - Contributing
 
 ## Data Structures
+
 The following data structures are currently implemented in this repo:
 
 - Linked List
@@ -18,16 +21,19 @@ The following data structures are currently implemented in this repo:
 - Heap
 - Graph
 - Hash Tables
+- Hash Maps
 - Trie
 - Array
 - ArrayList
- 
+
 Each data structure has its own folder containing its implementation, along with relevant tests.
 
 ## Algorithms
+
 The following algorithms are currently implemented in this repo:
 
 ### Sorting Algorithms:
+
 - Bubble Sort
 - Selection Sort
 - Insertion Sort
@@ -40,9 +46,11 @@ The following algorithms are currently implemented in this repo:
 Each algorithm has its own folder containing its implementation, along with relevant tests.
 
 ## Usage
+
 To use any of the implementations, clone this repo and import the relevant Java files into your project. Each implementation is self-contained in a single Java file, so it should be easy to use.
 
 To run the tests for any implementation, navigate to its folder and run the relevant test file. For example, to run the tests for the Linked List implementation, navigate to the "linked-list" folder and run the "LinkedListTest.java" file.
 
 ## Contributing
+
 I welcome contributions to this repo! If you find a bug or have an improvement to suggest, please feel free to open an issue or submit a pull request.
