+++ Stack/stack_methods.java
